Move `NthLastOverload` and `CollectOverloads` into overloads.d.ts
